requirements.txt
aiogram
environs
openai
python-dotenv